# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SofTeam Projects API",
        "description": "This API was made to handle Softeam project information",
        "version": "1.0.0"
    },
    "servers":[
        {
            "url": "http://localhost:3000",
            "description": "test server"
        }
    ],
    "paths": {
...
        "/project": {
            "get": {
                "summary": "Get all the projects",
                "description": "get all the projects that the Softeam is involved",
                "tags": ["Projects"],
                "security": [{"bearerAuth": []}],
                "responses": {
                    "500": {
                        "description": "Error trying to get projects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Error trying to connect with database"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type":"array",
                                    "items": {
                                        "$ref": "#/components/schemas/projects"
                                    }
                                },
                                "examples": {
                                    "projects": {
                                        "value": [
                                            {
                                                "_id": "projectID",
                                                "name": "Project",
                                                "description": "description 1",
                                                "status": "finished",
                                                "value": "R$3000,00",
                                                "collaborators": [
                                                    {
                                                        "_id":"collaboratorID",
                                                        "name": "collaborator",
                                                        "email": "collaborator@email.com",
                                                        "role": "Back-end developer"
                                                    }
                                                ],
                                                "companies": [
                                                    {
                                                        "_id": "companyID",
                                                        "name": "company",
                                                        "address": {
                                                            "country": "BR",
                                                            "state": "SP",
                                                            "city": "Sao Paulo",
                                                            "street": "Rua B",
                                                            "location": 150,
                                                            "zipCode": "70000-000"
                                                        },
                                                        "CNPJ": "11.111.111/1111-11",
                                                        "phone": "(11)91111-1111"
                                                    }
                                                ]
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/project/create": {
            "post": {
                "summary": "create one project",
                "description": "Create one project at the Softeam platform",
                "tags": ["Projects"],
                "security": [{"bearerAuth": []}],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/projects"
                            },
                            "examples": {
                                "project": {
                                    "value": {
                                        "_id": "projectID",
                                        "name": "Project",
                                        "description": "description 1",
                                        "status": "finished",
                                        "value": "R$3000,00",
                                        "collaborators": [
                                            {
                                                "_id":"collaboratorID",
                                                "name": "collaborator",
                                                "email": "collaborator@email.com",
                                                "role": "Back-end developer"
                                            }
                                        ],
                                        "companies": [
                                            {
                                                "_id": "companyID",
                                                "name": "company",
                                                "address": {
                                                    "country": "BR",
                                                    "state": "SP",
                                                    "city": "Sao Paulo",
                                                    "street": "Rua B",
                                                    "location": 150,
                                                    "zipCode": "70000-000"
                                                },
                                                "CNPJ": "11.111.111/1111-11",
                                                "phone": "(11)91111-1111"
                                            }
                                        ]   
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Insuficient Data!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error trying to create project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Error trying to connect with database"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "project Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/projects"
                                },
                                "examples": {
                                    "project": {
                                        "value": {
                                            "_id": "projectID",
                                            "name": "Project",
                                            "description": "description 1",
                                            "status": "finished",
                                            "value": "R$3000,00",
                                            "collaborators": [
                                                {
                                                    "_id":"collaboratorID",
                                                    "name": "collaborator",
                                                    "email": "collaborator@email.com",
                                                    "role": "Back-end developer"
                                                }
                                            ],
                                            "companies": [
                                                {
                                                    "_id": "companyID",
                                                    "name": "company",
                                                    "address": {
                                                        "country": "BR",
                                                        "state": "SP",
                                                        "city": "Sao Paulo",
                                                        "street": "Rua B",
                                                        "location": 150,
                                                        "zipCode": "70000-000"
                                                    },
                                                    "CNPJ": "11.111.111/1111-11",
                                                    "phone": "(11)91111-1111"
                                                }
                                            ]   
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/project/update/{_id}": {
            "put": {
                "summary": "Update one existing project",
                "description": "Update one existing project",
                "tags": ["Projects"],
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "description": "ID to find project",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Insuficient Data!"
                    },
                    "404": {
                        "description": "Company not found!"
                    },
                    "500": {
                        "description": "Error trying to update company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Error trying to connect with database"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/projects"
                                },
                                "examples": {
                                    "project": {
                                        "value": {
                                            "_id": "projectID",
                                            "name": "Project",
                                            "description": "description 1",
                                            "status": "finished",
                                            "value": "R$3000,00",
                                            "collaborators": [
                                                {
                                                    "_id":"collaboratorID",
                                                    "name": "collaborator",
                                                    "email": "collaborator@email.com",
                                                    "role": "Back-end developer"
                                                }
                                            ],
                                            "companies": [
                                                {
                                                    "_id": "companyID",
                                                    "name": "company",
                                                    "address": {
                                                        "country": "BR",
                                                        "state": "SP",
                                                        "city": "Sao Paulo",
                                                        "street": "Rua B",
                                                        "location": 150,
                                                        "zipCode": "70000-000"
                                                    },
                                                    "CNPJ": "11.111.111/1111-11",
                                                    "phone": "(11)91111-1111"
                                                }
                                            ]   
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/project/delete": {
            "put": {
                "summary": "Delete an existing project",
                "description": "Delete an existing project",
                "tags": ["Projects"],
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "_id",
                        "in": "query",
                        "description": "ID from the project that you want to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Insuficient Data!"
                    },
                    "500": {
                        "description": "Error trying to delete company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Error trying to connect with database"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/projects"
                                },
                                "examples": {
                                    "project": {
                                        "value": {
                                            "_id": "projectID",
                                            "name": "Project",
                                            "description": "description 1",
                                            "status": "finished",
                                            "value": "R$3000,00",
                                            "collaborators": [
                                                {
                                                    "_id":"collaboratorID",
                                                    "name": "collaborator",
                                                    "email": "collaborator@email.com",
                                                    "role": "Back-end developer"
                                                }
                                            ],
                                            "companies": [
                                                {
                                                    "_id": "companyID",
                                                    "name": "company",
                                                    "address": {
                                                        "country": "BR",
                                                        "state": "SP",
                                                        "city": "Sao Paulo",
                                                        "street": "Rua B",
                                                        "location": 150,
                                                        "zipCode": "70000-000"
                                                    },
                                                    "CNPJ": "11.111.111/1111-11",
                                                    "phone": "(11)91111-1111"
                                                }
                                            ]   
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
            "collaborators": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    }
                }
            },
            "companies": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "object",
                        "properties": {
                            "country": {
                                "type":"string"
                            },
                            "state": {
                                "type":"string"
                            },
                            "city": {
                                "type":"string"
                            },
                            "street": {
                                "type":"string"
                            },
                            "location": {
                                "type":"number"
                            },
                            "zipCode": {
                                "type":"string"
                            },
                            "complement": {
                                "type":"string"
                            }
                        }
                    },
                    "CNPJ": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    }
                }
            },
            "projects": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    },
                    "collaborators": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/collaborators"
                        }
                    },
                    "companies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/companies"
                        }
                    }
                }
            }
        },
        "securitySchemes" : {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}